refactored db.ts to smaller files
