--- mobile/src/database/db.ts
@@ -1,264 +1,0 @@
-import "react-native-get-random-values";
-import { type SQLiteDatabase } from "expo-sqlite";
-import { v4 as uuidv4 } from "uuid";
-import { DBLogItem, DBLogbook } from "@/src/types/db";
-
-export async function migrateDbIfNeeded(db: SQLiteDatabase) {
-	const DATABASE_VERSION = 4;
-
-	const versionRow = await db.getFirstAsync<{ user_version: number }>("PRAGMA user_version");
-	const currentDbVersion = versionRow?.user_version ?? 0;
-
-	console.log(`Current DB Version: ${currentDbVersion}. Target Version: ${DATABASE_VERSION}`);
-
-	if (currentDbVersion >= DATABASE_VERSION) {
-		console.log("Database is up to date.");
-
-		return;
-	}
-
-	console.log(`Starting Database Migration to v${DATABASE_VERSION}...`);
-
-	await db.execAsync("PRAGMA foreign_keys = ON");
-
-	await db.execAsync(`
-		DROP TABLE IF EXISTS vessels;
-		DROP TABLE IF EXISTS logs;
-	`);
-
-	await db.execAsync(`
-        CREATE TABLE IF NOT EXISTS logbooks (
-			id TEXT PRIMARY KEY NOT NULL,
-            name TEXT NOT NULL,
-            type TEXT,
-            registration TEXT,
-            owner_id TEXT,
-            created_at TEXT NOT NULL,
-            updated_at TEXT NOT NULL,
-            version INTEGER DEFAULT 0
-        );
-    `);
-
-	// Create Logs Table
-	await db.execAsync(`
-        CREATE TABLE IF NOT EXISTS log_items (
-			id TEXT PRIMARY KEY NOT NULL,
-            logbook_id TEXT NOT NULL,
-            title TEXT NOT NULL,
-            body TEXT,
-            latitude REAL NOT NULL,
-            longitude REAL NOT NULL,
-            created_at TEXT NOT NULL,
-            updated_at TEXT NOT NULL,
-            version INTEGER DEFAULT 0,
-            FOREIGN KEY (logbook_id) REFERENCES logbooks(id) ON DELETE CASCADE
-        );
-    `);
-
-	// Update version
-	await db.execAsync(`PRAGMA user_version = ${DATABASE_VERSION}`);
-	console.log("Database Migration Complete.");
-}
-
-// --- LOG FUNCTIONS ---
-
-export async function addLogItem(
-	db: SQLiteDatabase,
-	logbookId: string,
-	title: string,
-	body: string | null,
-	latitude: number,
-	longitude: number,
-): Promise<string> {
-	const id = uuidv4();
-	const now = new Date().toISOString();
-
-	await db.runAsync(
-		`INSERT INTO log_items (id, logbook_id, title, body, latitude, longitude, created_at, updated_at, version)
-         VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		id,
-		logbookId,
-		title,
-		body,
-		latitude,
-		longitude,
-		now, // created_at
-		now, // updated_at
-		0, // initial sync version
-	);
-	return id;
-}
-
-export async function getLogItems(db: SQLiteDatabase, logbookId?: string): Promise<DBLogItem[]> {
-	if (logbookId) {
-		return db.getAllAsync<DBLogItem>(`SELECT * FROM log_items WHERE logbook_id = ? ORDER BY created_at DESC`, logbookId);
-	}
-	return db.getAllAsync<DBLogItem>(`SELECT * FROM log_items ORDER BY created_at DESC`);
-}
-
-export async function getLogItemById(db: SQLiteDatabase, id: string): Promise<DBLogItem | null> {
-	return db.getFirstAsync<DBLogItem>(`SELECT * FROM log_items WHERE id = ?`, id);
-}
-
-export const updateLogItem = async (db: SQLiteDatabase, id: string, title: string, body: string | null, lat?: number, lon?: number) => {
-	const now = new Date().toISOString();
-
-	if (lat !== undefined && lon !== undefined) {
-		return await db.runAsync("UPDATE log_items SET title = ?, body = ?, latitude = ?, longitude = ?, updated_at = ? WHERE id = ?", [
-			title,
-			body,
-			lat,
-			lon,
-			now,
-			id,
-		]);
-	} else {
-		return await db.runAsync("UPDATE log_items SET title = ?, body = ?, updated_at = ? WHERE id = ?", [title, body, now, id]);
-	}
-};
-
-export async function deleteLogItem(db: SQLiteDatabase, id: string): Promise<void> {
-	await db.runAsync(`DELETE FROM log_items WHERE id = ?`, id);
-}
-
-// --- VESSEL FUNCTIONS ---
-
-/**
- * Inserts a new logbook (vessel) into the local SQLite db
- * * @param db - The active SQLite database context
- * @param id - The UUIDv4 generated by the remote PostgreSQL backend
- * @param name - The name of the vessel
- * @param type - The type of the vessel (retained locally)
- * @param registration - The registration number of the vessel (retained locally)
- * @returns The definitive ID of the created record
- */
-export async function addLogbook(db: SQLiteDatabase,ownerId: string, name: string, type: string, registration: string): Promise<string> {
-	const id = uuidv4();
-	const now = new Date().toISOString();
-	await db.runAsync(
-		`INSERT INTO logbooks (id, owner_id, name, type, registration, created_at, updated_at, version) VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
-		id,
-		ownerId,
-		name,
-		type,
-		registration,
-		now,
-		now,
-		0,
-	);
-	return id;
-}
-
-export async function getLogbooks(db: SQLiteDatabase, ownerId: string): Promise<DBLogbook[]> {
-	return db.getAllAsync<DBLogbook>(`SELECT * FROM logbooks WHERE owner_id = ? ORDER BY created_at DESC`, ownerId);
-}
-
-export async function deleteLogbook(db: SQLiteDatabase, id: string): Promise<void> {
-	await db.runAsync(`DELETE FROM logbooks WHERE id = ?`, id);
-}
-
-/**
- * Syncs an array of remote logbooks into the local SQLite db
- * Utilizes INSERT OR IGNORE to add new remote vessels without overwriting existing local ones
- * preventing Primary Key constraint violations during device migrations
- * @param db - The active SQLite database context
- * @param remoteLogbooks - The array of logbook DTOs returned from the Express API
- */
-export async function syncLogbooks(db: SQLiteDatabase, remoteLogbooks: any[], ownerId: string): Promise<void> {
-	for (const lb of remoteLogbooks) {
-		// Prisma returns 'vesselType' and 'registration', map them, falling back to empty strings if null.
-		const name = lb.name || "Unnamed Vessel";
-		const type = lb.vesselType || "";
-		const registration = lb.registration || "";
-
-		// Fallback to current time if backend timestamps are missing in the DTO
-		const createdAt = lb.createdAt || new Date().toISOString();
-		const updatedAt = lb.updatedAt || new Date().toISOString();
-		const version = lb.version || 0;
-
-		await db.runAsync(
-			`INSERT OR IGNORE INTO logbooks (id, owner_id, name, type, registration, created_at, updated_at, version)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?)`,
-			lb.id,
-			ownerId,
-			name,
-			type,
-			registration,
-			createdAt,
-			updatedAt,
-			version,
-		);
-	}
-}
-
-// --- MIGRATION FUNCTION ---
-
-/**
- * Takes any vessel currently marked as 'local' (offline user)
- * Re-assigns it to the newly logged-in user
- */
-export async function assignLocalDataToUser(db: SQLiteDatabase, userId: string): Promise<number> {
-	const result = await db.runAsync(
-		`UPDATE logbooks SET owner_id = ? WHERE owner_id = 'local' OR owner_id IS NULL`,
-		userId
-	);
-	// Returns the number of vessels that were moved to the new account
-	return result.changes;
-}
-
-/**
- * Finds all offline items that have not been successfully pushed to the server
- */
-export async function getPendingLogItems(db: SQLiteDatabase, logbookId: string) {
-	return db.getAllAsync<DBLogItem>(
-		`SELECT * FROM log_items WHERE logbook_id = ? AND version = 0`,
-		logbookId
-	);
-}
-
-/**
- * Finds the highest version number currently stored on this device for a specific logbook
- */
-export async function getHighestLogItemVersion(db: SQLiteDatabase, logbookId: string): Promise<number> {
-	const result = await db.getFirstAsync<{ max_version: number }>(
-		`SELECT MAX(version) as max_version FROM log_items WHERE logbook_id = ?`,
-		logbookId
-	);
-	return result?.max_version || 0;
-}
-
-/**
- * Intelligently merges remote server items into the local SQLite db
- */
-export async function mergeRemoteLogItems(db: SQLiteDatabase, logbookId: string, remoteItems: any[]): Promise<void> {
-	await db.withTransactionAsync(async () => {
-		for (const item of remoteItems) {
-			if (item.isActive === false) {
-				// wipe it locally if server says it is deleted
-				await db.runAsync(`DELETE FROM log_items WHERE id = ?`, item.id);
-			} else {
-				// insert the new log entry, or update an existing one
-				await db.runAsync(
-					`INSERT INTO log_items (id, logbook_id, title, body, latitude, longitude, created_at, updated_at, version)
-                     VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
-                     ON CONFLICT(id) DO UPDATE SET
-                     title = excluded.title,
-                     body = excluded.body,
-                     latitude = excluded.latitude,
-                     longitude = excluded.longitude,
-                     updated_at = excluded.updated_at,
-                     version = excluded.version`,
-					item.id,
-					logbookId,
-					item.title,
-					item.body || null, // Safely handle null descriptions
-					item.latitude,
-					item.longitude,
-					item.createdAt,
-					item.updatedAt,
-					item.version
-				);
-			}
-		}
-	});
-}
